fix: don't automatically import tf with determined
Previously we imported data_layer (which imports tf) with
the default train_context. We now import it only for the
contexts for estimators and tf.keras, which already have
tf imported.
