+++ harness/determined/__init__.py
@@ -1,6 +1,6 @@
 from determined.__version__ import __version__
 from determined._env_context import EnvContext
-from determined._train_context import _DataLayerContext, NativeContext, TrialContext
+from determined._train_context import NativeContext, TrialContext
 from determined._trial import Trial
 from determined._experiment_config import ExperimentConfig
 from determined._hparam import Categorical, Constant, Double, Integer, Log
