+++ harness/determined/_data_layer/__init__.py
@@ -1,0 +1,2 @@
+from determined._data_layer._data_layer import _CacheableDecorator
+from determined._data_layer._context import DataLayerContext
