+++ harness/determined/keras/__init__.py
@@ -15,7 +15,7 @@ from determined.keras._tf_keras_context import (
     TFKerasTrainConfig,
     TFKerasTrialContext,
 )
-from determined.keras._tf_keras_inputs import _init_input_manager
+from determined.keras._tf_keras_inputs import _init_input_managers
 from determined.keras._tf_keras_multi_gpu import _get_multi_gpu_model_and_optimizer
 from determined.keras._tf_keras_trial import TFKerasTrial, TFKerasTrialController
 from determined.keras._tf_keras_native import init
